feat: :art: moved fuctions into a seperate file for better code readability
